+++ resolwe_bio/tests/test_transmart.py
@@ -1,25 +1,23 @@
 # pylint: disable=missing-docstring
 import unittest
-
 from .utils import ProcessTestCase
 
 
 class TranSMARTProcessorTestCase(ProcessTestCase):
-
-    @unittest.skip("test data not ready")
+    @unittest.skip("processor connects to remote tranSMART server")
     def test_import(self):
         self.assertDataCount(0)
-        inputs = {'exps': 'transmart_log.xlsx'}
+        inputs = {'exps': 'transmart_log_exp.xlsx'}
         annotation = self.run_processor('import:web:transmart:expressions', inputs)
-        self.assertDataCount(7)
+        self.assertDataCount(5)
         self.assertFields(annotation, 'expset_type', 'Log2')
         self.assertFileExists(annotation, 'expset')
 
-    @unittest.skip("test data not ready")
+    @unittest.skip("processor connects to remote tranSMART server")
     def test_import_with_annotation(self):
         self.assertDataCount(0)
-        inputs = {'exps': 'transmart_log.xlsx', 'ann': 'transmart_log_annotation.xlsx'}
+        inputs = {'exps': 'transmart_log_exp.xlsx', 'ann': 'transmart_annotation.xlsx'}
         annotation = self.run_processor('import:web:transmart:expressions', inputs)
-        self.assertDataCount(7)
+        self.assertDataCount(5)
         self.assertFields(annotation, 'expset_type', 'Log2')
         self.assertFileExists(annotation, 'expset')